// src/app/models/collection-item.ts
export class CollectionItem {
    name: string = "Pikachu";
    description: string = "Pikachu is an Electric-type Pokémon introduced in Generation I. It evolves from Pichu when leveled up with high friendship and evolves into Raichu when exposed to a Thunder Stone.";
    price: number = 9.99;
    rarity: string = "Legendary";
    image: string = "img/pikachu.png";
}